components converted to tsx
