--- src/components/AuthContext.js
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-const AuthContext = createContext();
-
-export { AuthContext };
