--- src/components/AuthProvider.js
@@ -1,28 +1,0 @@
-import React, { useReducer, createContext } from "react";
-
-const AuthContext = createContext();
-
-const initialState = {
-  user: null,
-};
-
-const authReducer = (state, action) => {
-  switch (action.type) {
-    case "LOGIN":
-      return { ...state, user: action.payload };
-    default:
-      return state;
-  }
-};
-
-const AuthProvider = ({ children }) => {
-  const [state, dispatch] = useReducer(authReducer, initialState);
-
-  return (
-    <AuthContext.Provider value={{ state, dispatch }}>
-      {children}
-    </AuthContext.Provider>
-  );
-};
-
-export { AuthContext, AuthProvider };
